+++ CheeseQuest2/game/menu/LoadMenu.java
@@ -13,18 +13,25 @@ import game.system.*;
 public class LoadMenu extends Menu {
 
     public static final int MAXIMUM_SAVES = 5;
+    public static final String LOAD_GAME = "Load game";
+    public static final String NEW_GAME = "New game";
+    public static final String[] RETURN_TO_PREVIOUS_MENU = {"return","r"};
+    public static final String[] LOAD_OPTIONS = {"load","l"};
+    public static final String[] DELETE_OPTIONS = {"delete","d"};
     private static LoadMenu instance;
-    private ArrayList<String> validSave;
+    private ArrayList<String> validSaves;
+    private ArrayList<String> existingSaves;
+    private ArrayList<String> saveInfo;
 
     /**
      * Default LoadMenu constructor
      */
     private LoadMenu() {
-        existingSave = new ArrayList<String>();
-        validSave = new ArrayList<String>();
-        for (int i = 1; i <= MAXIMUM_SAVES; i++) {
-            validSave.add(Integer.toString(i));
-        }
+        existingSaves = new ArrayList<String>();
+        validSaves = new ArrayList<String>();
+        saveInfo = new ArrayList<String>();
+        determineValidSaves();
+        determineExistingSaves();
     }
 
     public static LoadMenu getInstance() {
@@ -40,28 +47,50 @@ public class LoadMenu extends Menu {
      */
     @Override
     public void outputPrompt() {
-        outputlnItem("Load Game");
+        determineExistingSaves();
+        outputlnTitle(LOAD_GAME);
         String gameName;
-        int i;
-        for (i = 0; i < validSave.size(); i++) {
-            if (existingSaves.contains(validSave.get(i))) {
-                gameName = "Load game"; // Game exists
-            } else {
-                gameName = "New game"; // Game does not exist
-            }
-            outputln("    " + i + ". " + gameName);
+        for (int i = 0; i < MAXIMUM_SAVES; i++) {
+            outputItem((i + 1) + ". ");
+            outputln(saveInfo.get(i));
         }
-        outputln("    " + i + ". New game");
+        outputln();
+        outputItem("Load");
+        output(" or ");
+        outputItem("Delete");
+        output(" a save file, or ");
+        outputItem("Return");
+        output(" to the ");
+        if (MenuManager.getInstance().getLastMenu().equals(MainMenu.getInstance())) {
+            outputln("Main Menu.");
+        } else {
+            outputln("the game.");
+        }
     }
 
     @Override
     public void processInput() {
-        if (verbEquals(existingSave)) { // Load preexiting game
-            loadGame(Integer.parseInt(getVerb()));
+        inputStartsWithStrip(LOAD_OPTIONS); // strip away load if added
+        if (inputStartsWith(existingSaves)) { // Load preexiting game
+            loadGame(Integer.parseInt(stripInput()));
             changeToGameMenu();
-        } else if (verbEquals(validSave)) { // Create new game
-            createGame(Integer.parseInt(verVerb()));
+        } else if (inputStartsWith(validSaves)) { // Create new game
+            createGame(Integer.parseInt(stripInput()));
+            loadGame(Integer.parseInt(stripInput())); // load game created
             changeToGameMenu();
+        } else if (inputStartsWithStrip(DELETE_OPTIONS)) {
+            if (inputStartsWith(existingSaves)) {
+                boolean deleted = WriteObject.getInstance().deleteSave();
+                if (deleted) {
+                    outputDeleted();
+                } else {
+                    outputNotDeleted();
+                }
+            } else {
+                outputNotDeleted();
+            }
+        } else if (inputStartsWithStrip(RETURN_TO_PREVIOUS_MENU)) {
+            changeToLastMenu();
         } else {
             outputInvalid();
             outputPrompt();
@@ -69,7 +98,15 @@ public class LoadMenu extends Menu {
     }
 
     public void outputInvalid() {
-        outputln("Choose a valid save file.");
+        outputln("Choose a save file from 1 to " + MAXIMUM_SAVES + ".");
+    }
+
+    public void outputDeleted() {
+        outputln("Save " + stripInput() + "deleted.");
+    }
+
+    public void outputNotDeleted() {
+        outputln("Save " + stripInput() + " does not exist and cannot be deleted.");
     }
 
 
@@ -78,8 +115,9 @@ public class LoadMenu extends Menu {
      * @param int saveNum of world to load
      */
     public void loadGame(int saveNum) {
-        ReadObject.setSaveNum(saveNum);
-        World world = (World) ReadObject.deserialize();
+        ReadObject reader = ReadObject.getInstance();
+        reader.setSaveNum(saveNum);
+        World world = (World) reader.deserialize();
         GameMenu gameMenu = GameMenu.getInstance();
         gameMenu.setWorld(world);
     }
@@ -89,10 +127,40 @@ public class LoadMenu extends Menu {
      * @param int saveNum of thworld to create
      */
     public void createGame(int saveNum) {
-        World world = new World();
-        WriteObject.setSaveNum(saveNum);
-        WriteObject.serialize(world);
-        GameMenu gameMenu = GameMenu.getInstance();
-        gameMenu.setWorld(world);
+        // World world = new World();
+        // WriteObject.setSaveNum(saveNum);
+        // WriteObject.serialize(world);
+        // GameMenu gameMenu = GameMenu.getInstance();
+        // gameMenu.setWorld(world);
+    }
+
+    /**
+     * Determines valid saves depending on MAXIMUM_SAVES
+     */
+    public void determineValidSaves() {
+        validSaves.clear();
+        for (int i = 1; i <= MAXIMUM_SAVES; i++) {
+            validSaves.add(Integer.toString(i));
+        }
+    }
+
+    /**
+     * [determineExistingSaves description]
+     */
+    public void determineExistingSaves() {
+        existingSaves.clear();
+        saveInfo.clear();
+        ReadObject reader = ReadObject.getInstance();
+        World world;
+        for (int i = 1; i <= MAXIMUM_SAVES; i++) {
+            reader.setSaveNum(i);
+            if (reader.saveExists()) {
+                world = (World) reader.deserialize();
+                existingSaves.add(Integer.toString(i));
+                saveInfo.add(world.getInfo());
+            } else {
+                saveInfo.add(NEW_GAME);
+            }
+        }
     }
 }