+++ CheeseQuest2/game/menu/Menu.java
@@ -14,7 +14,9 @@ public abstract class Menu {
     private static final String INPUT_MARKER = "> ";
     private String inputString;
     private String[] inputWords;
-    private String[] remainingWords;
+    // private String remainingString;
+    // private String[] remainingWords;
+    // private String verb;
     private OutputPanel outputPanel;
     private MenuManager menuManager; // Cannot have menuManeger, or infinite recursion occurs??
 
@@ -112,6 +114,87 @@ public abstract class Menu {
     public void outputlnRoom(String output) {
         outputRoom(output + "\n");
     }
+    public void outputTitle(String output) {
+        outputPanel.appendTitle(output);
+    }
+    public void outputlnTitle(String output) {
+        outputTitle(output + "\n");
+    }
+
+
+
+    // Input processing
+    // Words are processed one by one from the start and are stripped away
+
+
+    private boolean inputStartsWithChoice(ArrayList<String> arrayList, boolean strip) {
+        int longestWord = 0;
+        String[] words;
+        String word;
+
+        for (int i = 0; i < arrayList.size(); i++) {
+            words = arrayList.get(i).split(" "); // entries in array may have multiple words
+            if (words.length > longestWord) {
+                longestWord = words.length;
+                word = String.join(" ", Arrays.copyOfRange(inputWords,0,longestWord)); // make start comparison multiple words if need be
+                if (wordEquals(word,arrayList)) {
+                    if (strip) {
+                        stripInput(longestWord);
+                    }
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+    /**
+     * Checks if input starts with an element of arrayList
+     * If so, strips away starting words for inputWords and returns true
+     * @param  ArrayList<String> arrayList
+     * @return
+     */
+    public boolean inputStartsWithStrip(ArrayList<String> arrayList) {
+        return inputStartsWithChoice(arrayList, true);
+    }
+    public boolean inputStartsWithStrip(String[] array) {
+        return inputStartsWithStrip(new ArrayList<String>(Arrays.asList(array)));
+    }
+    public boolean inputStartsWithStrip(String word) {
+        return inputStartsWithStrip(new String[] {word});
+    }
+    // No strip versions
+    public boolean inputStartsWith(ArrayList<String> arrayList) {
+        return inputStartsWithChoice(arrayList, false);
+    }
+    public boolean inputStartsWith(String[] array) {
+        return inputStartsWith(new ArrayList<String>(Arrays.asList(array)));
+    }
+    public boolean inputStartsWith(String word) {
+        return inputStartsWith(new String[] {word});
+    }
+    /**
+     * Strips away starting words by index from inputWords
+     * If more words are to be stripped than exist, then all remaining words are stripped
+     * @param int index of new beginning inputWords
+     * @return words stripped away from input
+     */
+    public String stripInput(int index) {
+        String output = "";
+        if (inputWords.length < index) {
+            index = inputWords.length;
+        }
+        output = String.join(" ", Arrays.copyOfRange(inputWords,0,index));
+        inputWords = Arrays.copyOfRange(inputWords,index,inputWords.length);
+        return output;
+
+    }
+    /**
+     * Strips away 1 word from inputWords
+     * @return word stripped
+     */
+    public String stripInput() {
+        return stripInput(1);
+    }
 
     public boolean wordEquals(String word, String[] array) {
         return Arrays.asList(array).contains(word.toLowerCase());
@@ -131,15 +214,62 @@ public abstract class Menu {
     public boolean inputEquals(String word) {
         return wordEquals(getInputString(),word);
     }
-    public boolean verbEquals(String[] array) {
-        return wordEquals(getVerb(),array);
-    }
-    public boolean verbEquals(ArrayList<String> arrayList) {
-        return wordEquals(getVerb(),arrayList);
-    }
-    public boolean verbEquals(String word) {
-        return wordEquals(getVerb(),word);
-    }
+    /**
+     * Checks if inputString starts with an element of array
+     * set remaining words of inputString to remainingWords
+     * @param  String[] array         [description]
+     * @return          [description]
+     */
+    // public boolean verbEquals(String[] array) {
+    //     return verbEquals(new ArrayList<String>(Arrays.asList(array)));
+    //     // return wordEquals(getVerb(),array);
+    // }
+    /**
+     * Checks if input starts with an element of arrayList
+     * If so, sets starting word(s) to verb and rest to remainingWords
+     * @param  ArrayList<String> arrayList
+     * @return
+     */
+    // public boolean verbEquals(ArrayList<String> arrayList) {
+    //     int longestWord = 0;
+    //     String[] words;
+    //     String verb; // verb may be multiple words
+    //     for (int i = 0; i < arrayList.size(); i++) {
+    //         words = arrayList.get(i).split(" "); // entries in array may have multiple words
+    //         if (words.length > longestWord) {
+    //             longestWord = words.length;
+    //             verb = String.join(" ", Arrays.copyOfRange(inputWords,0,longestWord)); // make verb multiple words if need be
+    //             if (wordEquals(verb,arrayList)) {
+    //                 this.verb = verb;
+    //                 makeRemainingWords(longestWord);
+    //                 return true;
+    //             }
+    //         }
+    //     }
+    //     return false;
+    //     // return wordEquals(getVerb(),arrayList);
+    // }
+    // public boolean verbEquals(String word) {
+    //     return verbEquals(new String[] {word});
+    // }
+    /**
+     * From verb, set remaining words from inputWords to remainingWords
+     */
+    // public void makeRemainingWords(int start) {
+    //     remainingWords = Arrays.copyOfRange(inputWords,start,inputWords.length);
+    //     remainingString = String.join(" ",remainingWords);
+    // }
+
+    // public boolean remainingWordsEquals(){}
+    // public boolean remainingEquals(String[] array) {
+    //     return wordEquals(getRemainingString(),array);
+    // }
+    // public boolean remainingEquals(ArrayList<String> arrayList) {
+    //     return wordEquals(getRemainingString(),arrayList);
+    // }
+    // public boolean remainingEquals(String word) {
+    //     return wordEquals(getRemainingString(),word);
+    // }
 
     /**
     * Returns value of inputString
@@ -161,8 +291,16 @@ public abstract class Menu {
      * Returns first word of inputString
      * @return
      */
-    public String getVerb() {
-        return this.inputWords[VERB];
+    // public String getVerb() {
+    //     return this.verb;
+    // }
+
+    // public String[] getRemainingWords() {
+    //     return this.remainingWords;
+    // }
+    //
+    public String getRemainingString() {
+        return String.join(" ", inputWords);
     }
 
     /**
@@ -190,5 +328,8 @@ public abstract class Menu {
             System.out.println("Uh Oh...");
             e.printStackTrace();
         }
+    }
+    public void changeToLastMenu() {
+        menuManager.setMenu(menuManager.getLastMenu());
     }
 }