+++ CheeseQuest2/Changelog.md
@@ -6,6 +6,15 @@ Single-player text adventure game
 Developed in Java 1.8
 
 ## Changelog
+### 0.0.5 - September 1, 2017
+- Game driver CheeseQuest implemented and moved to main package
+- Inventory removes items if item count is non-positive
+- Person implemented
+- Title font added
+- Menu input processing changed
+    - No longer breaks input by verb and remaining words, as verb may be multiple words
+    - Now works similar to Cheese Quest 1, where input is processed by constantly stripping away words from the start of the inputString
+- Room not longer uses Coordinates class
 ### 0.0.4a - August 19, 2017
 - Changing between menus now works.
 ### 0.0.4 - August 18, 2017